# services/account/openapi.yaml
openapi: 3.0.3
info:
  title: EveryoneCoding 云端账号服务端 API
  description: >
    服务端复用既有账号认证，并提供发布版/目录/价格、平台钱包与账务对账 API。
    客户端可离线读取已保存目录快照；本服务不提供云同步、远程配置下发与分享链接。
  version: '1.7.0'
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000
    description: 本地开发
...
paths:
  /api/auth/register:
    post:
      tags: [auth]
      summary: 邮箱注册（自注册即开通）
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, description: '≥8 位且含两类字符（大写/小写/数字/特殊）' }
                displayName: { type: string }
      responses:
        '201':
          description: 注册成功，已开通默认工作区与免费权益包
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthPayload'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/auth/login:
    post:
      tags: [auth]
      summary: 邮箱 + 密码登录
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthPayload'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /api/auth/oauth/{provider}/authorize:
    get:
      tags: [auth]
      summary: 发起 OAuth 授权
      operationId: oauthAuthorize
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [wechat, google, github] }
        - name: redirect_uri
          in: query
          required: true
          schema: { type: string, format: uri }
        - name: code_challenge
          in: query
          required: true
          schema: { type: string, description: PKCE S256 code_challenge }
        - name: scope
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 返回授权 URL 与 state
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorizeUrl: { type: string }
                  state: { type: string }
        '400':
          $ref: '#/components/responses/Error'
  /api/auth/oauth/{provider}/callback:
    get:
      tags: [auth]
      summary: OAuth 回调换令牌（GET：浏览器重定向直连；POST：客户端转发回调参数）
      operationId: oauthCallback
      x-also-implemented-as-post: true
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [wechat, google, github] }
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
        - name: code_verifier
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 授权成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthResult'
        '400':
          $ref: '#/components/responses/Error'
  /api/auth/email/verify:
    post:
      tags: [auth]
      summary: 发送邮箱验证邮件（冷却窗口限流；未注册邮箱幂等成功）
      operationId: requestEmailVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: 已发送（或邮箱未注册/已验证）
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  emailVerified: { type: boolean, description: 仅已验证时返回 }
        '429':
          $ref: '#/components/responses/Error'
  /api/auth/email/verify/confirm:
    post:
      tags: [auth]
      summary: 确认邮箱验证（token 单次有效、过期拒绝）
      operationId: confirmEmailVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: 验证成功
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        '400':
          $ref: '#/components/responses/Error'

  # 邮件里链接指向的落地页：由服务自托管，不依赖桌面端是否运行
  # （渲染层用 HashRouter，裸路径 /verify-email 路由不到；且验证常在应用未启动时发生）
...
  /api/auth/email/status:
    get:
      tags: [auth]
      summary: 查询邮箱验证状态
      operationId: emailStatus
      security: []
      parameters:
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
      responses:
        '200':
          description: 验证状态
          content:
            application/json:
              schema:
                type: object
                properties: { emailVerified: { type: boolean } }
  /api/auth/password/reset/request:
    post:
      tags: [auth]
      summary: 请求重置密码验证码（6 位，冷却窗口限流；未注册邮箱幂等成功）
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: 已发送
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        '429':
          $ref: '#/components/responses/Error'
  /api/auth/password/reset:
    post:
      tags: [auth]
      summary: 验证码重置密码（码单次有效、过期拒绝；成功后撤销全部刷新令牌）
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, newPassword]
              properties:
                email: { type: string, format: email }
                code: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          description: 重置成功
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean } }
        '400':
          $ref: '#/components/responses/Error'
  /api/dev/email-outbox:
    get:
      tags: [auth]
      summary: 开发/运维读取邮件 outbox（收件人脱敏；生产建议 webhook 投递）
      operationId: devEmailOutbox
      security: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: 最近投递的邮件
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        to_addr: { type: string, description: 脱敏收件人 }
                        subject: { type: string }
                        body: { type: string }
                        kind: { type: string, enum: [verify, reset] }
                        created_at: { type: integer }
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: 刷新令牌（旧 refresh 轮换失效）
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: 新令牌对
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Error'
  /api/auth/bindings:
    get:
      tags: [auth]
      summary: 列出第三方绑定
      operationId: listBindings
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 绑定列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  bindings:
                    type: array
                    items: { $ref: '#/components/schemas/Binding' }
                  nextCursor: { type: string }
        '401':
          $ref: '#/components/responses/Error'
    post:
      tags: [auth]
      summary: 绑定第三方身份到当前账号（完整 OAuth 换身份流程）
      operationId: addBinding
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, code, state, codeVerifier]
              properties:
                provider: { type: string, enum: [wechat, google, github] }
                code: { type: string }
                state: { type: string }
                codeVerifier: { type: string }
      responses:
        '200':
          description: 绑定后的完整清单
          content:
            application/json:
              schema:
                type: object
                properties:
                  bindings:
                    type: array
                    items: { $ref: '#/components/schemas/Binding' }
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
    delete:
      tags: [auth]
      summary: 解绑第三方身份（按 bindingId；仅剩单一登录方式且无密码时 409）
      operationId: removeBinding
      security: [{ bearerAuth: [] }]
      parameters:
        - name: bindingId
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 解绑后的完整清单
          content:
            application/json:
              schema:
                type: object
                properties:
                  bindings:
                    type: array
                    items: { $ref: '#/components/schemas/Binding' }
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: 统一错误结构
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        traceId: { type: string }
...
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresAt: { type: integer, description: Access Token 过期毫秒时间戳 }
        refreshExpiresAt: { type: integer, description: Refresh Token 过期毫秒时间戳 }
    AccountIdentity:
      type: object
      properties:
        accountId: { type: string }
        login: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
        emailVerified: { type: boolean }
        hasPassword: { type: boolean }
    AuthPayload:
      type: object
      description: 注册/登录/OAuth 回调统一返回 { identity, tokens }
      properties:
        identity: { $ref: '#/components/schemas/AccountIdentity' }
        tokens: { $ref: '#/components/schemas/TokenPair' }
        workspaceId: { type: string }
        planId: { type: string }
        isNew: { type: boolean, description: 仅 OAuth 回调返回 }
    OAuthResult:
      allOf:
        - $ref: '#/components/schemas/AuthPayload'
    Binding:
      type: object
      properties:
        id: { type: string, description: 绑定行 id（解绑参数） }
        provider: { type: string }
        externalId: { type: string, description: 脱敏后的第三方标识 }
        boundAt: { type: integer }
...